+++ src/FsPickler/assemblyInfo.fs
@@ -1,9 +1,9 @@
 ﻿namespace System
 open System.Reflection
 
-[<assembly: AssemblyVersionAttribute("1.0.5")>]
-[<assembly: AssemblyFileVersionAttribute("1.0.5")>]
+[<assembly: AssemblyVersionAttribute("1.0.6")>]
+[<assembly: AssemblyFileVersionAttribute("1.0.6")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.0.5"
+    let [<Literal>] Version = "1.0.6"